# server/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tracking Service",
        "version": "1.0.0",
        "contact": {},
        "description": "share a tracking data with others, no authentication needed.\nthe data is represented by its hash, and will be kept only for 7 days."
    },
    "paths": {
		"/mobile-location": {
            "get": {
                "summary": "a socket-io server for gps mobile connect service",
				"tags": ["GPS Mobile Connect Service"],
                "description": "this service is used for creating live connection with mobile gps device.\n protocol is as follows:\n on connection the device 'role' is expected either 'recorder' or 'mobile'.\n role 'recorder':\n can only request to 'join' a room, in that case he will recive its room id in a 'code' message so he could share it with a mobile device.\n after joining a room he can listen to 'status' messages where he will recive a messages incicating if a moblie decive is connected or not(msg: mobile is/not connected).\n when a device is connected he can expect 'position' messages with the moblie gps coordinates.\n should listen on 'error' if anything go wrong in the process.\n role 'mobile':\n can 'join' a room with a code that was givin to the 'recorder', if successful he will be notified with a 'connected' message.\n after that he can send 'position' messages with its gps coordinates that will be sent to the 'recorder'.\n should listen on 'error' if anything go wrong in the process.\n",
                "parameters": [
					{
						"name": "role",
						"in": "query",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "either 'recorder' or 'moblie' based on what type of connection the user wants"
					}
				],
                "operationId": "MobileGpsTracking",
                "responses": {
                    "200": {
                        "description": "connected to service"
                    }
                }
            }
        },
        "/share-tracking/all": {
            "get": {
                "summary": "show all tracking",
				"tags": ["Tracking"],
                "description": "",
                "parameters": [],
                "operationId": "ShowAlllTracking",
                "responses": {
                    "200": {
                        "description": "Default response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Tracking_Record"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/share-tracking": {
            "get": {
                "summary": "get a tracking",
				"tags": ["Tracking"],
                "description": "",
                "parameters": [
                    {
                        "name": "hash",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Hash"
                        },
                        "description": ""
                    }
                ],
                "operationId": "GetTracking",
                "responses": {
                    "200": {
                        "description": "tracking data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Tracking_Record"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "callbacks": {}
            },
            "post": {
                "summary": "save a tracking",
				"tags": ["Tracking"],
                "description": "",
                "parameters": [],
                "operationId": "SaveTracking",
                "responses": {
                    "200": {
                        "description": "Default response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Hash"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid content"
                    }
                },
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Tracking_Record"
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "delete a tracking",
				"tags": ["Tracking"],
                "description": "",
                "parameters": [
                    {
                        "name": "hash",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Hash"
                        },
                        "description": ""
                    }
                ],
                "operationId": "DeleteTracking",
                "responses": {
                    "200": {
                        "description": "tracking was deleted successfully"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "callbacks": {}
            }
        }
    },
    "security": [],
    "servers": [
        {
            "url": "/",
            "description": "",
            "variables": {}
        }
    ],
    "components": {
        "links": {},
        "callbacks": {},
        "schemas": {
            "Tracking_Record": {
                "type": "object",
                "properties": {
                    "path": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Position"
                        }
                    },
                    "start_at": {
                        "type": "number"
                    },
                    "end_at": {
                        "type": "number"
                    },
                    "tag": {
                        "type": "string"
                    }
                }
            },
            "Position": {
                "type": "object",
                "properties": {
                    "latitude": {
                        "type": "number"
                    },
                    "longitude": {
                        "type": "number"
                    },
                    "time": {
                        "type": "number"
                    }
                }
            },
            "Hash": {
                "type": "object",
                "properties": {
                    "hash": {
                        "type": "string"
                    }
                }
            }
        },
        "securitySchemes": {}
    },
    "tags": []
}